pyproject.toml
[tool.poetry]
name = "auction-bot"
version = "0.1.0"
description = ""
authors = ["merllinsbeard"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.4.1"
asyncio = "^3.4.3"
pyyaml = "^6.0.1"
aiocryptopay = "^0.3.6"
aiohttp = "^3.9.3"
aiomysql = "^0.2.0"
tortoise-orm = "^0.20.0"
cryptography = "^42.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
